Add component-based grid system
A `Grid.js` module exports three different React components:
  `Container`, `Row` & `Columns`

Internally, these use a third-party Flexbox library.

The propTypes and/or backing library might get changed/replaced
when we field test them.

The `Grid.js` file is placed inside a `components` folder;
we'll see if this turns out to be a good way to structure the app.